refactor: move inner provider classes into separate files
