+++ provider-androidx/src/main/java/io/github/iprodigy/cache/provider/androidx/AndroidLruProvider.java
@@ -5,10 +5,7 @@ import io.github.iprodigy.cache.api.Cache;
 import io.github.iprodigy.cache.api.ICacheSpec;
 import io.github.iprodigy.cache.api.RemovalListener;
 import io.github.iprodigy.cache.api.domain.RemovalCause;
-import io.github.iprodigy.cache.core.AbstractCache;
 import io.github.iprodigy.cache.core.AbstractCacheProvider;
-import lombok.EqualsAndHashCode;
-import lombok.Value;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -37,41 +34,5 @@ public final class AndroidLruProvider extends AbstractCacheProvider {
 					listener.onRemoval(key, oldValue, cause);
 			}
 		};
-	}
-
-	@Value
-	@EqualsAndHashCode(callSuper = false)
-	static class LruDelegate<K, V> extends AbstractCache<K, V> {
-		LruCache<K, V> cache;
-
-		@Override
-		public V get(K key) {
-			return cache.get(key);
-		}
-
-		@Override
-		public V put(K key, V value) {
-			return cache.put(key, value);
-		}
-
-		@Override
-		public V remove(K key) {
-			return cache.remove(key);
-		}
-
-		@Override
-		public void clear() {
-			cache.evictAll();
-		}
-
-		@Override
-		public long size() {
-			return cache.size();
-		}
-
-		@Override
-		protected Object getLock() {
-			return this.cache;
-		}
 	}
 }
