+++ provider-caffeine2/src/main/java/io/github/iprodigy/cache/provider/caffeine2/Caffeine2Provider.java
@@ -7,11 +7,6 @@ import io.github.iprodigy.cache.api.ICacheSpec;
 import io.github.iprodigy.cache.api.domain.ExpiryType;
 import io.github.iprodigy.cache.api.domain.RemovalCause;
 import io.github.iprodigy.cache.core.AbstractCacheProvider;
-import io.github.iprodigy.cache.core.delegate.GenericMapCacheDelegate;
-import lombok.EqualsAndHashCode;
-import lombok.Value;
-
-import java.util.function.Function;
 
 public final class Caffeine2Provider extends AbstractCacheProvider {
 
@@ -45,38 +40,6 @@ public final class Caffeine2Provider extends AbstractCacheProvider {
 			case COLLECTED:
 			default:
 				return RemovalCause.OTHER;
-		}
-	}
-
-	@Value
-	@EqualsAndHashCode(callSuper = false)
-	private static class CaffeineDelegate<K, V> extends GenericMapCacheDelegate<K, V> {
-		com.github.benmanes.caffeine.cache.Cache<K, V> cache;
-
-		public CaffeineDelegate(com.github.benmanes.caffeine.cache.Cache<K, V> cache) {
-			super(cache.asMap());
-			this.cache = cache;
-		}
-
-		@Override
-		public V get(K key) {
-			return cache.getIfPresent(key);
-		}
-
-		@Override
-		public V computeIfAbsent(K key, Function<K, V> computeFunc) {
-			return cache.get(key, computeFunc);
-		}
-
-		@Override
-		public void clear() {
-			cache.invalidateAll();
-		}
-
-		@Override
-		public long size() {
-			cache.cleanUp();
-			return cache.estimatedSize();
 		}
 	}
 }
