+++ provider-infinispan/src/main/java/io/github/iprodigy/cache/provider/infinispan/InfinispanDelegate.java
@@ -1,0 +1,52 @@
+package io.github.iprodigy.cache.provider.infinispan;
+
+import io.github.iprodigy.cache.api.Cache;
+import lombok.Value;
+
+import java.util.function.BiFunction;
+import java.util.function.Function;
+
+@Value
+class InfinispanDelegate<K, V> implements Cache<K, V> {
+    org.infinispan.Cache<K, V> cache;
+
+    @Override
+    public V get(K key) {
+        return cache.get(key);
+    }
+
+    @Override
+    public V put(K key, V value) {
+        return cache.put(key, value);
+    }
+
+    @Override
+    public V remove(K key) {
+        return cache.remove(key);
+    }
+
+    @Override
+    public void clear() {
+        cache.clear();
+    }
+
+    @Override
+    public long size() {
+        return cache.size();
+    }
+
+    @Override
+    public V computeIfAbsent(K key, Function<K, V> computeFunc) {
+        return cache.computeIfAbsent(key, computeFunc);
+    }
+
+    @Override
+    public V putIfAbsent(K key, V value) {
+        return cache.putIfAbsent(key, value);
+    }
+
+    @Override
+    public V merge(K key, V value, BiFunction<V, V, V> mergeFunc) {
+        return cache.merge(key, value, mergeFunc);
+    }
+}
